fix(kotlin-sdk): address PR #4284 review findings + CI
CI: rustfmt the JNI descriptor table (the Rust-workspace job failed on
cargo fmt --check); the PR title uppercase-subject violation was fixed
on the PR itself.

Review findings addressed:

- Claim submits ONLY the URI its displayed preview validated
  (previewedUri gate) — editing the field after a valid parse can no
  longer submit an unvalidated URI; the preview also re-parses when the
  claim wallet appears (deep-link seeding raced the wallet flows).
- Scanned invitation links never touch SavedStateHandle: a one-shot
  in-memory scanResultSink on AppUiState carries bearer scans (the
  scanner clears it on delivery AND on cancel so a stale sink can't
  hijack a later generic scan); the parked pendingInviteUri is now
  released when the claim starts or the sheet is explicitly closed,
  not at seeding — surviving activity recreation in between.
- Identity HD-slot selection for claim + reclaim-register unions Room
  rows with slots held by in-flight RegistrationCoordinator entries.
- The create-result stage sets FLAG_SECURE while the bearer QR/link is
  on screen and holds the host dismissal gate until Done (an accidental
  scrim-tap discarded the only shareable copy); didCopy resets when the
  delayed clipboard clear actually fires; the reclaim sheet's
  no-manager early return releases the busy gate it had just taken.
- nowUnix carried as Long end-to-end with a 1..=u32::MAX guard on both
  invitation JNI entry points (the Int default would go negative in
  2038 while the shared u32 C API stays valid); a successful create
  that returns no link now throws (voucher recorded + reclaimable)
  instead of yielding "".
- Parity manifest: dropped the duplicate asset-lock-resume verification
  entry (merge artifact); summary regenerated. TEST_PLAN DP-15 no
  longer calls the amount floor an expiry bound.

JVM suites + :app:assembleDebug green; cargo fmt --check clean; JNI lib
tests 37/37.

+++ packages/kotlin-sdk/sdk/src/main/kotlin/org/dashfoundation/dashsdk/ffi/DashpayNative.kt
@@ -229,7 +229,7 @@ internal object DashpayNative {
         fundingAccountIndex: Int,
         inviterIdentityId: ByteArray?,
         inviterUsername: String?,
-        nowUnix: Int,
+        nowUnix: Long,
         coreSignerHandle: Long,
     ): String?
 
@@ -248,6 +248,6 @@ internal object DashpayNative {
         identityIndex: Int,
         pubkeysBlob: ByteArray,
         signerHandle: Long,
-        nowUnix: Int,
+        nowUnix: Long,
     ): IdentityRegistrationNativeResult
 }
